Fixed your formatting a bit, read description
nice unc :relaxed:  you're using swing I like it.

try and stick to a format in your code on indentation etc, so it is all consistent. look at my edits and you'll see what I mean

Also if you are overriding implement or extend methods you should use the @Override annotation to show that you are overriding an existing method from a superclass.

This is good though I like it :+1: